requirements.txt
pybind11
fasttext
textblob
nltk
pandas
pyspellchecker